requirements.txt
pytz==2019.3
waitress==1.4.1
Flask==1.1.1
# Note on confluent_kafka. for avro to work you'll need to edit Python\Lib\site-packages\confluent_kafka\schema_registry_client\avro.py
# line 321 should be obj_dict = schemaless_reader(payload, writer_schema, None)
# Link to github issue: https://github.com/confluentinc/confluent-kafka-python/issues/887
confluent_kafka==1.5.0
pyOpenSSL==19.1.0
PyYAML==5.3.1
requests==2.25.0
fastavro==1.2.0